+++ routes/youtubers.js
@@ -1,27 +1,24 @@
 const express = require("express");
 const router = express.Router();
-const dbFunc = require("./func/dbFunc");
-
+const dbFunc = require('../func/dbfunc');
+const dbdata = require('../data/data');
 router.use(express.json());
 
 
 router
     .route('/')
     .post((req, res) => {// 유튜버 등록록
-
         const newYoutuber = req.body;
         console.log(newYoutuber);
         if (newYoutuber.nickname != undefined) {
-            // console.log(db); // test
-            // console.log(db.has(newYoutuber.nickname));
-            if (!db.has(newYoutuber.nickname)) {
+            if (!dbdata.db.has(newYoutuber.nickname)) {
                 let newId = 0;
-                db.forEach((value, key) => {
+                dbdata.db.forEach((value, key) => {
                     if (value.id > newId) {
                         newId = value.id;
                     }
                 });
-                db.set(newYoutuber.nickname, {
+                dbdata.db.set(newYoutuber.nickname, {
                     id: newId + 1,
                     userId: newYoutuber.userId,
                     pwd: newYoutuber.pwd,
@@ -30,7 +27,6 @@ router
                     desc: newYoutuber.desc,
                     subscribers: 0
                 });
-                // console.log(db); //test
                 res.status(201).json({ "message": `${newYoutuber.nickname}님, 새로운 유투버로 등록이 완료됬습니다.` });
             }
             else (
@@ -42,9 +38,9 @@ router
         }
     })
     .get((req, res) => {//유튜버 전체 조회
-        if (db.size !== 0) {
-            let videos = Object.fromEntries(videoDB);
-            let youtubers = Object.fromEntries(db);
+        if (dbdata.db.size !== 0) {
+            let videos = Object.fromEntries(dbdata.videoDB);
+            let youtubers = Object.fromEntries(dbdata.db);
             res.status(200).json({ "message": "전체 유튜버 및 영상 조회입니다.", "youtubers": youtubers, "videos": videos });
 
         }
@@ -54,8 +50,8 @@ router
         }
     })
     .delete((req, res) => {
-        if (db.size > 0) {
-            db.clear();
+        if (dbdata.db.size > 0) {
+            dbdata.db.clear();
             res.status(200).json({ "message": "모든 유튜버를 제거했습니다" })
         } else {
             res.status(404).json({ "message": "이미 비어 있습니다" });
@@ -66,8 +62,8 @@ router
 
 router.delete('/:id', (req, res) => {
     const id = parseInt(req.params.id);
-    if (db.has(id)) {
-        let data = db.get(id);
+    if (dbdata.db.has(id)) {
+        let data = dbdata.db.get(id);
         db.delete(id);
         res.status(200).json({ "message": `${data.nickname}님을 제거했습니다` });
     } else {
@@ -81,7 +77,7 @@ router.post('/login', (req, res) => {
     let youtubeUser;
 
     // 유튜버 체크. Id, Pwd가 매치됬는지 확인하고, 매치되었다면 유튜버 정보 저장
-    for (let youtuber of db.values()) {
+    for (let youtuber of dbdata.db.values()) {
         if (youtuber.userId === userId) {
             console.log("id was matched");
             isIdMatched = true;
